--- src/pages/TasksPage/db.json
@@ -1,151 +1,0 @@
-{
-  "tasks": {
-    "items": [
-      {
-        "sprintId": "PikRSLReU2HGoin2hzGm",
-        "title": "Первая задача",
-        "hoursPlanned": "12",
-        "hoursWasted": 9,
-        "hoursWastedPerDay": [
-          {
-            "currentDay": "06.04.2021",
-            "singleHoursWasted": 7
-          },
-          {
-            "currentDay": "07.04.2021",
-            "singleHoursWasted": 2
-          },
-          {
-            "currentDay": "08.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "09.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "12.04.2021",
-            "singleHoursWasted": 0
-          }
-        ],
-        "id":"Xf31JGeVUKQtpxQOsWmG"
-      },
-      {
-        "sprintId": "PikRSLReU2HGoin2hzGm",
-        "title":"Вторая задача",
-        "hoursPlanned": "8",
-        "hoursWasted": 5,
-        "hoursWastedPerDay": [
-          {
-            "currentDay": "06.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "07.04.2021",
-            "singleHoursWasted": 5
-          },
-          {
-            "currentDay": "08.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "09.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "12.04.2021",
-            "singleHoursWasted": 0
-          }
-        ],
-        "id":"hKaAycEaw9dcEGW2aYFd"
-      },
-      {
-        "sprintId": "PikRSLReU2HGoin2hzGm",
-        "title":"Третья задача",
-        "hoursPlanned": "10",
-        "hoursWasted": 7,
-        "hoursWastedPerDay": [
-          {
-            "currentDay": "06.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "07.04.2021",
-            "singleHoursWasted": 3
-          },
-          {
-            "currentDay": "08.04.2021",
-            "singleHoursWasted": 4
-          },
-          {
-            "currentDay": "09.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "12.04.2021",
-            "singleHoursWasted": 0
-          }
-        ],
-        "id":"s4rQpGo44AVVuRDaW0l0"
-      },
-      {
-        "sprintId": "PikRSLReU2HGoin2hzGm",
-        "title":"Четвертая задача",
-        "hoursPlanned": "4",
-        "hoursWasted": 4,
-        "hoursWastedPerDay": [
-          {
-            "currentDay": "06.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "07.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "08.04.2021",
-            "singleHoursWasted": 4
-          },
-          {
-            "currentDay": "09.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "12.04.2021",
-            "singleHoursWasted": 0
-          }
-        ],
-        "id":"NoKtBX0dHmqMFlFUyryU"
-      },
-      {
-        "sprintId": "PikRSLReU2HGoin2hzGm",
-        "title":"Пятая задача",
-        "hoursPlanned": "6",
-        "hoursWasted": 6,
-        "hoursWastedPerDay": [
-          {
-            "currentDay": "06.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "07.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "08.04.2021",
-            "singleHoursWasted": 0
-          },
-          {
-            "currentDay": "09.04.2021",
-            "singleHoursWasted": 5
-          },
-          {
-            "currentDay": "12.04.2021",
-            "singleHoursWasted": 1
-          }
-        ],
-        "id":"fcDg57rMyVb4TEWOFKXP"
-      }
-    ]
-  }
-}
